requirements.txt
streamlit>=1.32.0
streamlit-folium>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
folium>=0.15.0
plotly>=5.18.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
geopy>=2.4.0
streamlit-lottie>=0.0.5
